refactor(base): improve organization of z-offset API
