+++ packages/@sanity/base/src/components/index.ts
@@ -1,4 +1,4 @@
 export * from './formField'
 export * from './transitional'
 export * from './UserAvatar'
-export * from './ZIndexProvider'
+export * from './zOffsets'
