+++ packages/@sanity/base/src/components/zOffsets/index.ts
@@ -1,0 +1,2 @@
+export * from './ZIndexProvider'
+export * from './useZIndex'
